requirements.txt
## These requirements were autogenerated by pipenv# To regenerate from the project's Pipfile, run:##    pipenv lock --requirements#-i https://pypi.org/simpleaniso8601==9.0.1bcrypt==3.2.0; python_version >= '3.6'blinker==1.4cffi==1.14.5click==8.0.1; python_version >= '3.6'colorama==0.4.4; platform_system == 'Windows'dnspython==2.1.0; python_version >= '3.6'email-validator==1.1.3flask-bcrypt==0.7.1flask-jwt-extended==4.2.1flask-mail==0.9.1flask-mongoengine==1.0.0flask-restful==0.3.9flask-wtf==0.15.1; python_version >= '3.6'flask==2.0.1gunicorn==20.1.0idna==3.2; python_version >= '3.5'itsdangerous==2.0.1; python_version >= '3.6'jinja2==3.0.1; python_version >= '3.6'markupsafe==2.0.1; python_version >= '3.6'mongoengine==0.23.1; python_version >= '3.6'pycparser==2.20; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'pyjwt==2.1.0; python_version >= '3.6'pymongo==3.11.4pytz==2021.1six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'werkzeug==2.0.1; python_version >= '3.6'wtforms[email]==2.3.3python-decouple==3.4